report-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
kafka-python==2.0.2
structlog==23.2.0
prometheus-client==0.19.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0 